# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: AREA Server API
  version: 0.1.0
  description: |-
    The AREA Server API powers the automation platform that links actions and reactions across
    partner services. Use it to onboard users, issue secure sessions, and orchestrate OAuth-based
    integrations.

    ### Highlights
    - Account lifecycle endpoints for registration, verification, and session management.
    - OAuth helpers to initiate and complete third-party connector authorisations.
    - Metadata routes that expose the current backend build, feature flags, and compatibility notes.

    ### Getting started
    1. Create a user via `POST /v1/users` and complete the verification journey.
    2. Authenticate with either password credentials or OAuth connectors to obtain an `area_session` cookie.
    3. Call authenticated endpoints such as `GET /v1/auth/me` with the session cookie attached as a standard cookie header.

    Refer to the environment selector in the sidebar to toggle between production, staging, and local URLs.
    
    ### Rate limits & reliability
    - Default rate limit: **120 requests per minute** per authenticated session.
    - Burst tolerance: short spikes up to 300 requests per minute are accepted for 30 seconds.
    - All endpoints are reachable over HTTPS only; staging and local URLs are for testing purposes.

    ### Support & SLA
    - Incident response via the on-call rotation is available 24/7 through the status portal.
...
  contact:
...
  license:
...
  x-logo:
...
servers:
...
security:
  - sessionAuth: []
paths:
...
  /v1/users:
    post:
      summary: Register a new user
      description: |-
        Begins the email-first account creation workflow. Upon success the backend sends a
        verification message and returns a tracker ID so clients can surface progress updates.
      operationId: registerUser
      tags:
        - auth
      security: []
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -X POST https://api.area-server.app/v1/users \
              -H 'Content-Type: application/json' \
              -d '{"email":"alex.martin@example.com","password":"Passw0rd!23"}'
        - lang: javascript
          label: JavaScript (fetch)
          source: |
            await fetch('https://api.area-server.app/v1/users', {
              method: 'POST',
              headers: { 'Content-Type': 'application/json' },
              body: JSON.stringify({
                email: 'alex.martin@example.com',
                password: 'Passw0rd!23',
              }),
            });
        - lang: python
          label: Python (requests)
          source: |
            import requests

            payload = {
                "email": "alex.martin@example.com",
                "password": "Passw0rd!23",
            }
            response = requests.post(
                'https://api.area-server.app/v1/users', json=payload, timeout=10
            )
            response.raise_for_status()
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
            examples:
              standard:
                summary: Email-first registration
                value:
                  email: alex.martin@example.com
                  password: Passw0rd!23
      responses:
        '202':
          description: Registration accepted, verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterUserResponse'
              examples:
                accepted:
                  summary: Registration acknowledged
                  value:
                    userId: usr_01H9ZQF8E4XKC5V7QMMC1782X9
        '400':
          description: Invalid input
        '409':
          description: Email already registered
  /v1/auth/verify:
    post:
      summary: Verify email address and create a session
      description: |-
        Confirms the registration token from the verification email. Upon success the API issues
        an authenticated session and returns its metadata for client storage.
      operationId: verifyEmail
      tags:
        - auth
      security: []
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -X POST https://api.area-server.app/v1/auth/verify \
              -H 'Content-Type: application/json' \
              -d '{"token":"01H9ZQF8E4XKC5V7QMMC1782X9"}' \
              -c cookies.txt
        - lang: javascript
          label: JavaScript (fetch)
          source: |
            await fetch('https://api.area-server.app/v1/auth/verify', {
              method: 'POST',
              headers: { 'Content-Type': 'application/json' },
              credentials: 'include',
              body: JSON.stringify({ token: '01H9ZQF8E4XKC5V7QMMC1782X9' }),
            });
        - lang: python
          label: Python (requests)
          source: |
            import requests

            payload = {"token": "01H9ZQF8E4XKC5V7QMMC1782X9"}
            session = requests.Session()
            response = session.post(
                'https://api.area-server.app/v1/auth/verify', json=payload, timeout=10
            )
            response.raise_for_status()
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
            examples:
              verification:
                summary: Verify a new registration
                value:
                  token: 01H9ZQF8E4XKC5V7QMMC1782X9
      responses:
        '200':
          description: Email verified and session issued
          headers:
            Set-Cookie:
              description: Authentication session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
              examples:
                session:
                  summary: Session metadata
                  value:
                    tokenType: session
                    expiresAt: 2024-04-18T11:32:00Z
        '400':
          description: Invalid token
        '410':
          description: Token expired or already used
...
  /v1/auth/me:
    get:
      summary: Retrieve the authenticated user profile
      description: |-
        Returns the canonical user profile associated with the supplied `area_session` cookie.
        Useful for bootstrapping client state after reloads.
      operationId: getCurrentUser
      tags:
        - auth
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl https://api.area-server.app/v1/auth/me -b cookies.txt
        - lang: javascript
          label: JavaScript (fetch)
          source: |
            const response = await fetch('https://api.area-server.app/v1/auth/me', {
              credentials: 'include',
            });
            const currentUser = await response.json();
        - lang: python
          label: Python (requests)
          source: |
            import requests

            session = requests.Session()
            response = session.get('https://api.area-server.app/v1/auth/me', timeout=10)
            response.raise_for_status()
            print(response.json())
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                profile:
                  summary: Authenticated user
                  value:
                    id: usr_01H9ZQF8E4XKC5V7QMMC1782X9
                    email: alex.martin@example.com
        '401':
          description: Session missing or expired
  /v1/identities:
    get:
      summary: List connected identities
      description: Retrieve OAuth identities linked to the authenticated user
      operationId: listIdentities
      tags:
        - auth
      responses:
        '200':
          description: Linked identities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityListResponse'
        '401':
          description: Authentication required
        '500':
          description: Failed to list identities
  /v1/oauth/{provider}/authorize:
    post:
      summary: Initiate OAuth authorization
      description: |-
        Creates a provider-specific authorisation URL and state token. Invoke this endpoint before
        redirecting users to complete OAuth consent flows.
      operationId: authorizeOAuth
      tags:
        - auth
      parameters:
        - $ref: '#/components/parameters/OAuthProvider'
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -X POST https://api.area-server.app/v1/oauth/google/authorize \
              -H 'Content-Type: application/json' \
              -d '{"redirectUri":"https://app.area-server.app/oauth/complete"}'
        - lang: javascript
          label: JavaScript (fetch)
          source: |
            const response = await fetch('https://api.area-server.app/v1/oauth/google/authorize', {
              method: 'POST',
              headers: { 'Content-Type': 'application/json' },
              credentials: 'include',
              body: JSON.stringify({
                redirectUri: 'https://app.area-server.app/oauth/complete',
              }),
            });
            const { authorizationUrl } = await response.json();
        - lang: python
          label: Python (requests)
          source: |
            import requests

            payload = {
                "redirectUri": "https://app.area-server.app/oauth/complete",
            }
            response = requests.post(
                'https://api.area-server.app/v1/oauth/google/authorize',
                json=payload,
                timeout=10,
            )
            response.raise_for_status()
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthAuthorizationRequest'
            examples:
              override-redirect:
                summary: Customise the redirect URI
                value:
                  redirectUri: https://app.area-server.app/oauth/complete
      responses:
        '200':
          description: Authorization parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthAuthorizationResponse'
              examples:
                jump-url:
                  summary: Consent URL to open in browser
                  value:
                    authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth?client_id=...
        '400':
          description: Invalid request
        '404':
          description: Provider not configured
        '501':
          description: OAuth not configured
  /v1/oauth/{provider}/exchange:
    post:
      summary: Exchange authorization code for session
      description: |-
        Completes the OAuth handshake by trading the provider-issued code for an AREA session. The
        backend validates the state token implicitly through the `area_session` cookie.
      operationId: exchangeOAuth
      tags:
        - auth
      parameters:
        - $ref: '#/components/parameters/OAuthProvider'
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -X POST https://api.area-server.app/v1/oauth/google/exchange \
              -H 'Content-Type: application/json' \
              -d '{"code":"4/0AbCDefGhIjKlMnOpQrStUvWxYz"}' \
              -b cookies.txt -c cookies.txt
        - lang: javascript
          label: JavaScript (fetch)
          source: |
            await fetch('https://api.area-server.app/v1/oauth/google/exchange', {
              method: 'POST',
              headers: { 'Content-Type': 'application/json' },
              credentials: 'include',
              body: JSON.stringify({ code: '4/0AbCDefGhIjKlMnOpQrStUvWxYz' }),
            });
        - lang: python
          label: Python (requests)
          source: |
            import requests

            payload = {"code": "4/0AbCDefGhIjKlMnOpQrStUvWxYz"}
            session = requests.Session()
            response = session.post(
                'https://api.area-server.app/v1/oauth/google/exchange',
                json=payload,
                timeout=10,
            )
            response.raise_for_status()
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthExchangeRequest'
            examples:
              standard:
                summary: Exchange the provider code
                value:
                  code: 4/0AbCDefGhIjKlMnOpQrStUvWxYz
      responses:
        '200':
          description: Authentication succeeded
          headers:
            Set-Cookie:
              description: Authentication session cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
              examples:
                session:
                  summary: Session metadata
                  value:
                    tokenType: session
                    expiresAt: 2024-04-18T11:32:00Z
        '400':
          description: Invalid payload
        '404':
          description: Provider not configured
        '501':
          description: OAuth not configured
        '502':
          description: Provider exchange failed
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: area_session
      description: Session cookie issued after authentication.
  schemas:
...
    RegisterUserRequest:
      type: object
      description: Payload used to enrol a new AREA account prior to email verification.
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: Primary email address that receives activation and security alerts.
          example: alex.martin@example.com
        password:
          type: string
          format: password
          minLength: 8
          description: Strong password that satisfies backend policy (min. 8 chars, letters + numbers).
          example: Passw0rd!23
      example:
        email: alex.martin@example.com
        password: Passw0rd!23
    RegisterUserResponse:
      type: object
      description: Confirmation payload indicating the asynchronous verification workflow started.
      required: [userId, expiresAt]
      properties:
        userId:
          type: string
          description: Unique identifier assigned to the provisioned user.
          example: usr_01H9ZQF8E4XKC5V7QMMC1782X9
        expiresAt:
          type: string
          format: date-time
          description: Expiration timestamp of the verification token.
          example: 2024-04-18T11:32:00Z
    VerifyEmailRequest:
      type: object
      description: Token issued in the verification email used to activate the account.
      required: [token]
      properties:
        token:
          type: string
          description: Single-use verification token issued as part of registration.
          example: 01H9ZQF8E4XKC5V7QMMC1782X9
      example:
        token: 01H9ZQF8E4XKC5V7QMMC1782X9
    AuthSessionResponse:
      type: object
      description: Session descriptor mirroring the cookie issued by the backend.
      required: [user]
      properties:
        tokenType:
          type: string
          example: session
          description: Grants semantics of the issued token; currently `session`.
        expiresAt:
          type: string
          format: date-time
          description: Expiration timestamp (UTC) aligning with the `area_session` cookie expiry.
        user:
          $ref: '#/components/schemas/User'
      example:
        tokenType: session
        expiresAt: 2024-04-18T11:32:00Z
        user:
          id: usr_01H9ZQF8E4XKC5V7QMMC1782X9
          email: alex.martin@example.com
          status: active
          createdAt: 2024-04-01T10:00:00Z
          updatedAt: 2024-04-01T10:00:00Z
...
    UserResponse:
      type: object
      description: Public representation of the authenticated user.
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      description: Detailed user payload returned by authenticated endpoints.
      required: [id, email, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier assigned to the user.
        email:
          type: string
          format: email
          description: Primary email associated with the account.
        status:
          type: string
          description: Lifecycle status (`pending`, `active`, etc.).
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp in UTC.
        updatedAt:
          type: string
          format: date-time
          description: Timestamp of the latest account update in UTC.
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the most recent successful login.
    OAuthAuthorizationRequest:
      type: object
      description: Optional parameters forwarded to the configured OAuth provider.
      properties:
        redirectUri:
          type: string
          format: uri
          description: Explicit redirect URI to override the default callback registered with the provider.
          example: https://app.area-server.app/oauth/complete
        state:
          type: string
          description: Opaque value used to maintain state between the request and callback.
        prompt:
          type: string
          description: Provider-specific prompt parameter allowing the user experience to be tweaked.
        usePkce:
          type: boolean
          description: Flag indicating whether PKCE should be enforced.
        scopes:
          type: array
          description: Additional scopes to request from the provider.
          items:
            type: string
    OAuthAuthorizationResponse:
      type: object
      description: Provider authorisation metadata.
      required: [authorizationUrl]
      properties:
        authorizationUrl:
          type: string
          format: uri
          description: URL that the client must open to continue authorisation with the provider.
          example: https://accounts.google.com/o/oauth2/v2/auth?client_id=...
        state:
          type: string
          description: State echoed by the provider after authorisation.
        codeVerifier:
          type: string
          description: PKCE code verifier to persist until the exchange completes.
        codeChallenge:
          type: string
          description: PKCE code challenge derived from the verifier.
        codeChallengeMethod:
          type: string
          description: Method used to compute the PKCE code challenge.
    OAuthExchangeRequest:
      type: object
      description: Payload received after the provider redirect to complete the exchange.
      required: [code]
      properties:
        code:
          type: string
          description: One-time authorisation code issued by the provider.
          example: 4/0AbCDefGhIjKlMnOpQrStUvWxYz
        redirectUri:
          type: string
          format: uri
          description: Redirect URI supplied during the authorisation flow.
        codeVerifier:
          type: string
          description: PKCE code verifier required when the provider enforces PKCE.
        state:
          type: string
          description: Provider-issued state parameter used to prevent CSRF.
...
    IdentityListResponse:
      type: object
      required: [identities]
      properties:
        identities:
          type: array
          items:
            $ref: '#/components/schemas/IdentitySummary'
    IdentitySummary:
      type: object
      required: [id, provider, subject, connectedAt]
      properties:
        id:
          type: string
          format: uuid
        provider:
          type: string
        subject:
          type: string
        scopes:
          type: array
          items:
            type: string
        connectedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
  parameters:
    OAuthProvider:
      name: provider
      in: path
      required: true
      description: Slug corresponding to a configured OAuth provider (for example `google`, `github`).
      schema:
        type: string
      example: google